# backend/cmd/main/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Petplace API Description",
        "title": "Petplace API Version1",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:5000",
    "basePath": "/api",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "log in user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Log In",
                "parameters": [
                    {
                        "description": "Login payload",
                        "name": "LoginPayload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.LoginPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/auth/signup": {
            "post": {
                "description": "sign up a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Sign Up",
                "parameters": [
                    {
                        "description": "Signup payload",
                        "name": "SignUpPayload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.User": {
            "type": "object",
            "required": [
                "email",
                "first_name",
                "password",
                "surename"
            ],
            "properties": {
                "age": {
                    "type": "integer"
                },
                "citizen_id": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "expiry": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image_profile": {
                    "type": "string"
                },
                "name": {
                    "description": "credit card",
                    "type": "string"
                },
                "number": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "paypal_email": {
                    "type": "string"
                },
                "security_code": {
                    "type": "string"
                },
                "surename": {
                    "type": "string"
                },
                "tel": {
                    "type": "string"
                }
            }
        },
...
        "types.LoginPayload": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}